+++ examples/04-topography/4.3-plot_topographic_diffusion.py
@@ -4,7 +4,7 @@ Topographic Diffusion
 
 .. rubric:: By Dennise Valadez and David Minton
 
-This example demonstrates how to use the :meth:`apply_diffusion` method in the :ref:`Surface <ug-surface>` class to model topographic diffusion. In this example, we will simulate the change in elevation over time of a hill with a Gaussian profile. This example has an analytical solution, which we will compare against the numerical solution provided by the diffusion method.
+This example demonstrates how to use the :py:meth:`apply_diffusion` method in the :ref:`Surface <ug-surface>` class to model topographic diffusion. In this example, we will simulate the change in elevation over time of a hill with a Gaussian profile. This example has an analytical solution, which we will compare against the numerical solution provided by the diffusion method.
 
 For this example we will use the "hireslocal" surface type, which will give us a high-resolution local region of the surface that will be approximately a flat surface.